Forgot swithing frames last time

+++ videopreviewcreator.h
@@ -17,8 +17,8 @@ public:
     QByteArray getImageData(fas_raw_image_type *);
     QImage getPreview(QString);
     QVector<int> getHistogram(QImage);
-    QVector<int> getMedian(QVector<QPair<QVector<int>, int> >);
-    QPair<QVector<int>, int> getClosest(QVector<QPair<QVector<int>, int> > hists, QVector<int> median);
+    QVector<qint64> getMedian(QVector<QPair<QVector<int>, int> >);
+    QPair<QVector<int>, int> getClosest(QVector<QPair<QVector<int>, int> > hists, QVector<qint64> median);
 };
 
 #endif // VIDEOPREVIEWCREATOR_H
